Add MRT file parse.  We can try it with 'cargo run updates.20240301.0000'.

+++ src/packet/attr.rs
@@ -78,7 +78,7 @@ pub struct AtomicAggregateAttr {}
 
 #[derive(Debug, NomBE)]
 pub struct AggregatorAttr {
-    pub asn: u16,
+    pub asn: u32,
     pub ip: u32,
 }
